모임 - issue 제거됨 / 참가 신청 수락,취소,출석 체크

+++ src/main/java/com/studyolle/event/EventService.java
@@ -50,8 +50,26 @@ public class EventService {
 
 	public void cancelEnrollment(Event event, Account account) {
 		Enrollment enrollment = enrollmentRepository.findByEventAndAccount(event, account);
-		event.removeEnrollment(enrollment);
-		enrollmentRepository.delete(enrollment); // 참가 신청 삭제
-		event.acceptNextWaitingEnrollment(); // 선착순의 경우 참가대기 있는 경우 자동 참가 처리
+		if (!enrollment.isAttended()) {
+			event.removeEnrollment(enrollment);
+			enrollmentRepository.delete(enrollment);
+			event.acceptNextWaitingEnrollment();
+		}
+	}
+
+	public void acceptEnrollment(Event event, Enrollment enrollment) {
+		event.accept(enrollment);
+	}
+
+	public void rejectEnrollment(Event event, Enrollment enrollment) {
+		event.reject(enrollment);
+	}
+
+	public void checkInEnrollment(Enrollment enrollment) {
+		enrollment.setAttended(true);
+	}
+
+	public void cancelCheckInEnrollment(Enrollment enrollment) {
+		enrollment.setAttended(false);
 	}
 }
